requirements.txt
pandas
SQLAlchemy
PyMySQL
google.cloud.storage
#mysql-connector-python

